requirements.txt
fastapi
uvicorn[standard]
python-multipart
itsdangerous
Jinja2
starlette
httpx
requests
pydantic
pydantic-settings

# Google OAuth and APIs
google-auth
google-auth-oauthlib
google-api-python-client

# RAG stack
faiss-cpu
sentence-transformers
numpy
scikit-learn
chromadb

# Optional LLM providers (previous OpenAI/Groq kept optional)
openai
groq
# Gemini (Google Generative AI)
google-generativeai

# Streamlit UI
streamlit

# PDF parsing
pdfminer.six
